Chore: Move development packages to devDependencies
This commit moves several packages that are only used during development from the dependencies section to the devDependencies section of package.json. These packages include type definitions and linting tools which are not needed in production.

By doing this, we reduce the size of our production build and potentially decrease build and deployment times.